requirements.txt
appdirs==1.4.0
astroid==2.1.0
diff-match-patch==20110725.1
dj-database-url==0.5.0
Django==2.2.4
django-excel==0.0.6
django-heroku==0.3.1
django-import-export==1.1.0
et-xmlfile==1.0.1
flake8==3.3.0
gunicorn==19.6.0
isort==4.3.4
jdcal==1.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
odfpy==1.3.4
openpyxl==2.4.8
packaging==16.8
psycopg2==2.8.3
psycopg2-binary==2.7.6.1
pycodestyle==2.3.1
pyexcel==0.4.5
pyexcel-io==0.3.4.1
pyexcel-ods==0.3.3
pyexcel-webio==0.0.11
pyexcel-xls==0.3.3
pyexcel-xlsx==0.3.0
pyflakes==1.5.0
pylint==2.2.2
pyparsing==2.1.10
pytz==2017.2
PyYAML==4.2b1
six==1.10.0
sqlparse==0.3.0
tablib==0.12.1
texttable==0.9.0
unicodecsv==0.14.1
virtualenv==15.1.0
whitenoise==3.2
wrapt==1.10.11
xlrd==1.0.0
xlwt==1.2.0